sperate tool from main
